+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Nonfunctional
+
+- Migrate the project structure to use Poetry. [issue-34](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/34)
+
 ## [2.0.3] - 2021-09-14
 
 ### Fixed