+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `Tier` property in the SSM parameter section. [issue-27](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/27)
+- Add support for AWS Encription SDK. [issue-28](https://github.com/lucasvieirasilva/aws-ssm-secrets-cli/issues/28)
 
 ### Changed